re-orgaize file structure
